requirements.txt
google-cloud-bigquery
requests
pandas
db-dtypes